fix(ui): change id to numeric ids

+++ app/api/middlewares/auth.go
@@ -53,7 +53,7 @@ func AuthMiddleware(c *gin.Context) {
 	ctx = context.WithValue(ctx, internal.ContextIdentity, userID)
 	// fill request context with values here
 
-	c.Set("userID", userID)
+	c.Set("userID", uint(userID))
 	c.Request = c.Request.WithContext(ctx)
 	c.Next()
 }
